Using previous/next pages
In this commit, the previous button will lead to the previous page, and the next button will lead to the next page. No longer do you need to hard-code which page each button leads to!
Storing mPageId so that it can be used in StoryActivity where we check if the page is the first page - this does not have to be used, you could also just use the local choice variable. Added getters/setters for this.
Refactored mFinalPage to mSinglePage, so that the single button can be used on either the first page or last pages.